feat: create, update 통합해서 Member, ArtWork 이벤트 리팩토링
