+++ src/main/java/com/sptp/backend/art_work/event/ArtWorkEvent.java
@@ -1,5 +1,6 @@
-package com.sptp.backend.member.event;
+package com.sptp.backend.art_work.event;
 
+import com.sptp.backend.art_work.repository.ArtWork;
 import com.sptp.backend.common.NotificationCode;
 import com.sptp.backend.member.repository.Member;
 import lombok.Getter;
@@ -7,8 +8,9 @@ import lombok.RequiredArgsConstructor;
 
 @Getter
 @RequiredArgsConstructor
-public class MemberUpdateEvent {
+public class ArtWorkEvent {
 
     private final Member member;
+    private final ArtWork artwork;
     private final NotificationCode notificationCode;
 }
