+++ src/main/java/com/sptp/backend/art_work/event/ArtWorkEventListener.java
@@ -2,6 +2,9 @@ package com.sptp.backend.art_work.event;
 
 import com.sptp.backend.art_work.repository.ArtWork;
 import com.sptp.backend.common.NotificationCode;
+import com.sptp.backend.member.repository.Member;
+import com.sptp.backend.notification.repository.Notification;
+import com.sptp.backend.notification.repository.NotificationRepository;
 import lombok.RequiredArgsConstructor;
 import lombok.extern.slf4j.Slf4j;
 import org.springframework.context.event.EventListener;
@@ -16,13 +19,29 @@ import org.springframework.transaction.annotation.Transactional;
 @RequiredArgsConstructor
 public class ArtWorkEventListener {
 
+    private final NotificationRepository notificationRepository;
+
+    // 작품 등록 완료 알림, 경매 등록 알림, 전시회 등록 알림, 낙찰 알림
     @EventListener
-    public void handleArtWorkCreateEvent(ArtWorkCreateEvent artworkCreateEvent){
+    public void handleArtWorkEvent(ArtWorkEvent artWorkEvent){
 
-        ArtWork artwork = artworkCreateEvent.getArtwork();
-        NotificationCode notificationCode = artworkCreateEvent.getNotificationCode();
-        //log.info(NotificationCode.addArtWorkTitle(artwork.getTitle(), notificationCode));
+        Member member = artWorkEvent.getMember();
+        ArtWork artWork = artWorkEvent.getArtwork();
+        NotificationCode notificationCode = artWorkEvent.getNotificationCode();
 
+        saveNotification(member, artWork, notificationCode);
+    }
 
+    public void saveNotification(Member member, ArtWork artWork, NotificationCode notificationCode){
+
+        Notification notification = Notification.builder()
+                .member(member)
+                .title(notificationCode.getTitle())
+                .message(artWork.getTitle() + notificationCode.getTitle())
+                .details(notificationCode.getDetails())
+                .link(notificationCode.getLink())
+                .build();
+
+        notificationRepository.save(notification);
     }
 }
