+++ src/main/java/com/sptp/backend/member/event/MemberEvent.java
@@ -1,6 +1,5 @@
-package com.sptp.backend.art_work.event;
+package com.sptp.backend.member.event;
 
-import com.sptp.backend.art_work.repository.ArtWork;
 import com.sptp.backend.common.NotificationCode;
 import com.sptp.backend.member.repository.Member;
 import lombok.Getter;
@@ -8,9 +7,8 @@ import lombok.RequiredArgsConstructor;
 
 @Getter
 @RequiredArgsConstructor
-public class ArtWorkCreateEvent {
+public class MemberEvent {
 
     private final Member member;
-    private final ArtWork artwork;
     private final NotificationCode notificationCode;
 }
